refactor(event-agenda): move locale formatters to date-time-helper
Extracts LOCALE_FORMATTERS and applyLocaleFormat from the event-agenda
block into date-time-helper.js so they can be reused by session-hub and
other blocks without duplication.
